feat: Update bookinstance model with formatted due date
This commit modifies the `bookinstance.js` file to add a virtual property `due_back_formatted` to the `BookInstanceSchema`. This property uses the Luxon library to format the `due_back` date in a human-readable format. The formatted due date is then used in the `bookinstance_list.ejs` view to display the due date of book instances that are not available.