feat: Update author list view with formatted dates
This commit modifies the `author_list.ejs` view to display the formatted dates of birth and death for each author. It adds virtual properties `date_of_birth_formatted` and `date_of_death_formatted` to the `AuthorSchema` in the `author.js` model. These properties use the Luxon library to format the dates in a human-readable format. The formatted dates are then used in the view to provide a more user-friendly representation of the author's birth and death dates.