feat: Update genre list view and controller
This commit modifies the `genreController.js` file to fetch all genres and render them in the `genre_list.ejs` view. The view is also created in this commit. It displays a list of genres with links to their detail pages. The genres are fetched from the database and sorted alphabetically by name.